:bookmark: reformatted code

+++ src/main/java/com/luna/common/net/IPAddressUtil.java
@@ -60,7 +60,7 @@ public class IPAddressUtil {
                     }
 
                     var2 *= 10L;
-                    var2 += (long)var9;
+                    var2 += var9;
                     var5 = false;
                 }
             }
